requirements.txt
dj-database-url==0.4.2
Django==1.8.18
django-debug-toolbar==1.5
djangorestframework==3.6.4
gunicorn==19.4.5
psycopg2==2.6.1
sqlparse==0.2.3
whitenoise==3.0
